chore: rename htmlElements file
